CR amends - move uplink controller's file
