+++ src/Common/Libraries/Provider.php
@@ -28,7 +28,7 @@ class Provider extends Service_Provider {
 		$this->container->singleton( static::class, $this );
 
 		tribe_register_provider( Installer\Provider::class );
-		tribe_register_provider( Uplink\Controller::class );
+		tribe_register_provider( Uplink_Controller::class );
 
 		DB\Config::setHookPrefix( $this->get_hook_prefix() );
 		Assets\Config::set_hook_prefix( $this->get_hook_prefix() );
